src/main/java/yandex.practicum/http/HttpTaskServer.java
package yandex.practicum.http;

import com.sun.net.httpserver.HttpServer;
import yandex.practicum.tracker.service.ITaskManager;
import yandex.practicum.tracker.service.Managers;

import java.io.IOException;
import java.net.InetSocketAddress;

public class HttpTaskServer {
    public HttpServer httpServer;
    private static final int PORT = 8080;
    private static final ITaskManager taskManager = Managers.getDefault();

    public HttpTaskServer() throws IOException {
        httpServer = HttpServer.create();
        httpServer.bind(new InetSocketAddress(PORT), 0);
        httpServer.createContext("/tasks", new TasksHandler(taskManager));
        httpServer.createContext("/epics", new EpicsHandler(taskManager));
        httpServer.createContext("/subtasks", new SubtasksHandler(taskManager));
        httpServer.createContext("/history", new HistoryHandler(taskManager));
        httpServer.createContext("/prioritized", new PrioritizedHandler(taskManager));
    }

    public void start() {
        httpServer.start();
    }

    public void stop() {
        httpServer.stop(0);
    }
}
